+++ examples/cifar100_mobilenetv2.py
@@ -3,56 +3,10 @@ from torch import nn
 from torchvision import transforms, datasets
 from torch.utils.data import DataLoader
 
-from tqdm import tqdm
-
 from bitnet.models.mobilenet_v2 import mobilenet_v2, bit_mobilenet_v2
 from bitnet.metrics import Metrics
 from bitnet.seed import set_seed
-
-
-device = torch.device('cuda') if torch.cuda.is_available() else 'cpu'
-
-
-def train_model(
-        model: nn.Module,
-        train_loader: DataLoader,
-        optimizer: torch.optim.Optimizer,
-        criterion: nn.CrossEntropyLoss,
-        num_epochs: int) -> None:
-    for epoch in range(num_epochs):
-        pbar = tqdm(total=len(train_loader), desc=f"Training {model.__name__}")
-        running_loss: float = 0.0
-        for i, (images, labels) in enumerate(train_loader):
-            images, labels = images.to(device), labels.to(device)
-            outputs = model(images)
-            loss = criterion(outputs, labels)
-
-            optimizer.zero_grad()
-            loss.backward()
-            optimizer.step()
-
-            running_loss += loss.item()
-            pbar.set_description(
-                f'Model: {model.__name__} - Epoch [{epoch+1}], Loss: {running_loss / (i+1):.4f}'
-            )
-            pbar.update(1)
-        pbar.close()
-
-
-def test_model(model: nn.Module, test_loader: DataLoader) -> tuple[dict[str, float], Metrics]:
-    metrics_used: Metrics = Metrics.ACCURACY
-    model.eval()
-    correct: int = 0
-    total: int = 0
-    with torch.no_grad():
-        for images, labels in test_loader:
-            images, labels = images.to(device), labels.to(device)
-            outputs = model(images)
-            _, predicted = torch.max(outputs.data, 1)
-            total += labels.size(0)
-            correct += (predicted == labels).sum().item()
-    metric: float = 100 * correct / total
-    return {model.__name__: metric}, metrics_used
+from bitnet.base_experiment import train_model, test_model
 
 
 def run(seed: int | None) -> tuple[dict[str, float], Metrics, int]:
@@ -70,9 +24,8 @@ def run(seed: int | None) -> tuple[dict[str, float], Metrics, int]:
         transforms.Normalize([0.5071, 0.4865, 0.4409], [0.2623, 0.2513, 0.2714])
     ])
 
-    print(f"Testing on {device=}")
-    bitnet =  bit_mobilenet_v2(num_classes=num_classes, pretrained=True).to(device)
-    floatnet = mobilenet_v2(num_classes, pretrained=True).to(device)
+    bitnet =  bit_mobilenet_v2(num_classes=num_classes, pretrained=True)
+    floatnet = mobilenet_v2(num_classes, pretrained=True)
     num_params_bitnet: int = sum(p.numel() for p in bitnet.parameters() if p.requires_grad)
     num_params_floatnet: int = sum(p.numel() for p in floatnet.parameters() if p.requires_grad)
     assert num_params_bitnet == num_params_floatnet
