+++ examples/mnist_feedforward.py
@@ -1,86 +1,13 @@
-from collections.abc import Callable
 import torch
-from torch import Tensor
 from torch import nn
 from torchvision import transforms, datasets
 from torch.utils.data import DataLoader
 
-from tqdm import tqdm
-
 from bitnet.nn.bitlinear import BitLinear
 from bitnet.metrics import Metrics
 from bitnet.seed import set_seed
-
-
-device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-
-
-class Net(nn.Module):
-    def __init__(
-            self,
-            linear_layer: Callable,
-            input_size: int,
-            hidden_size: int,
-            num_classes: int
-        ) -> None:
-        super(Net, self).__init__()
-        self.linear_layer = linear_layer
-        self.fc1 = linear_layer(input_size, hidden_size)
-        self.fc2 = linear_layer(hidden_size, num_classes)
-
-
-    @property
-    def __name__(self) -> str:
-        return "BitNet" if self.linear_layer == BitLinear else "FloatNet"
-
-
-    def forward(self, _input: Tensor) -> Tensor:
-        out = _input.flatten(start_dim=1)
-        out = torch.relu(self.fc1(out))
-        out = self.fc2(out)
-        return out
-
-
-def train_model(
-        model: nn.Module,
-        train_loader: DataLoader,
-        optimizer: torch.optim.Optimizer,
-        criterion: nn.CrossEntropyLoss,
-        num_epochs: int) -> None:
-    for epoch in range(num_epochs):
-        pbar = tqdm(total=len(train_loader), desc=f"Training {model.__name__}")
-        running_loss: float = 0.0
-        for i, (images, labels) in enumerate(train_loader):
-            images, labels = images.to(device), labels.to(device)
-            outputs = model(images)
-            loss = criterion(outputs, labels)
-
-            optimizer.zero_grad()
-            loss.backward()
-            optimizer.step()
-
-            running_loss += loss.item()
-            pbar.set_description(
-                f'Model: {model.__name__} - Epoch [{epoch+1}], Loss: {running_loss / (i+1):.4f}'
-            )
-            pbar.update(1)
-        pbar.close()
-
-
-def test_model(model: nn.Module, test_loader: DataLoader) -> tuple[dict[str, float], Metrics]:
-    metrics_used: Metrics = Metrics.ACCURACY
-    model.eval()
-    correct: int = 0
-    total: int = 0
-    with torch.no_grad():
-        for images, labels in test_loader:
-            images, labels = images.to(device), labels.to(device)
-            outputs = model(images)
-            _, predicted = torch.max(outputs.data, 1)
-            total += labels.size(0)
-            correct += (predicted == labels).sum().item()
-    metric: float = 100 * correct / total
-    return {model.__name__: metric}, metrics_used
+from bitnet.models.feedforward import Feedforward
+from bitnet.base_experiment import train_model, test_model
 
 
 def run(seed: int | None) -> tuple[dict[str, float], Metrics, int]:
@@ -100,9 +27,8 @@ def run(seed: int | None) -> tuple[dict[str, float], Metrics, int]:
         transforms.Normalize((0.1307,), (0.3081,))
     ])
 
-    print(f"Testing on {device=}")
-    bitnet = Net(BitLinear, input_size, hidden_size, num_classes).to(device)
-    floatnet = Net(nn.Linear, input_size, hidden_size, num_classes).to(device)
+    bitnet = Feedforward(BitLinear, input_size, hidden_size, num_classes)
+    floatnet = Feedforward(nn.Linear, input_size, hidden_size, num_classes)
     num_params_bitnet: int = sum(p.numel() for p in bitnet.parameters() if p.requires_grad)
     num_params_floatnet: int = sum(p.numel() for p in floatnet.parameters() if p.requires_grad)
     assert num_params_bitnet == num_params_floatnet
